mcp-server/apigee-mcp-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
requests==2.31.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
fastmcp==0.1.0
pyyaml==6.0.1 